noita-proxy: socketaddr binding via --host
This allows binding to arbitrary IP addresses (`[2001:db8::1]:12345`, `203.0.113.1:12345`) while still providing compatibility to the previous behaviour of binding to a port on IPv4 only.
The only non-compatible change is that when providing neither a valid `SocketAddr` nor a valid port the code will now err out instead of defaulting to a default port to avoid silently discarding arguments or invalid configurations.
As a result `--host --exe-path` will now fail as `--exe-path` is not a valid address or port.

Minor changes:

- some practically infallible parsing (`Ipv4Addr::UNSPECIFIED`) has also been replaced by infallible typing avoiding unnecessary `unwrap()`s
- typos

Fixes #389

Signed-off-by: benaryorg <binary@benary.org>